VMC/circledetect/python/requirements.txt
loguru==0.6.0
numpy==1.23.1
bell-avr-libraries[mqtt]==0.1.9
transforms3d==0.3.1
opencv-python

